+++ packages/sst/src/bootstrap.ts
@@ -197,9 +197,10 @@ export async function bootstrapSST() {
       path.resolve(__dirname, "support/bootstrap-metadata-function")
     ),
     handler: "index.handler",
-    runtime: (region?.startsWith("us-gov-") || region?.startsWith("cn-"))
-      ? Runtime.NODEJS_16_X
-      : Runtime.NODEJS_18_X,
+    runtime:
+      region?.startsWith("us-gov-") || region?.startsWith("cn-")
+        ? Runtime.NODEJS_16_X
+        : Runtime.NODEJS_18_X,
     environment: {
       BUCKET_NAME: bucket.bucketName,
     },
@@ -270,7 +271,6 @@ async function bootstrapCDK() {
   const identity = await useSTSIdentity();
   const credentials = await useAWSCredentials();
   const { region, profile, cdk } = useProject().config;
-  cdk || {};
   await new Promise<void>((resolve, reject) => {
     const proc = spawn(
       [
@@ -290,6 +290,9 @@ async function bootstrapCDK() {
         ...(cdk?.qualifier ? ["--qualifier", cdk.qualifier] : []),
         ...(cdk?.fileAssetsBucketName
           ? ["--toolkit-bucket-name", cdk.fileAssetsBucketName]
+          : []),
+        ...(cdk?.customPermissionsBoundary
+          ? ["--custom-permissions-boundary", cdk.customPermissionsBoundary]
           : []),
       ].join(" "),
       {