+++ packages/sst/src/cli/sst.ts
@@ -16,7 +16,7 @@ dotenv.config({
   override: true,
 });
 
-import { env } from "./commands/env.js";
+import { bootstrap } from "./commands/bootstrap.js";
 import { dev } from "./commands/dev.js";
 import { bind } from "./commands/bind.js";
 import { build } from "./commands/build.js";
@@ -30,11 +30,11 @@ import { diff } from "./commands/diff.js";
 import { version } from "./commands/version.js";
 import { telemetry } from "./commands/telemetry.js";
 
+bootstrap(program);
 dev(program);
 deploy(program);
 build(program);
 bind(program);
-env(program);
 secrets(program);
 remove(program);
 update(program);