+++ packages/sst/test/constructs/remix-site/sst-env.d.ts
@@ -1,0 +1,1 @@
+/// <reference path="../../../../../../../../../var/folders/0_/wf8r61p114vg8d5g3jysb4cr0000gn/T/sst/0808d738-dd0b-4ae7-8f2c-ed33d64b2efe/.sst/types/index.ts" />